e2etest: move scylla_cluster from validator-tests
This commit moves interface of scylla_cluster actor from validator-test into
e2etest-scylla_cluster. It also updates crate references.
