+++ crates/validator/src/auth.rs
@@ -7,12 +7,12 @@ use crate::TestActors;
 use crate::common;
 use crate::common::*;
 use async_backtrace::framed;
+use e2etest_scylla_cluster::ScyllaClusterExt;
 use std::sync::LazyLock;
 use std::time::Duration;
 use tokio::time::sleep;
 use tracing::info;
 use uuid::Uuid;
-use vector_search_validator_tests::ScyllaClusterExt;
 use vector_search_validator_tests::TestCase;
 use vector_search_validator_tests::VectorStoreClusterExt;
 use vector_store::httproutes::IndexStatus;
