+++ crates/validator/src/high_availability.rs
@@ -8,10 +8,10 @@ use std::collections::HashMap;
 use crate::TestActors;
 use crate::common::*;
 use async_backtrace::framed;
+use e2etest_scylla_cluster::ScyllaClusterExt;
+use e2etest_scylla_cluster::ScyllaNodeConfig;
 use scylla::statement::Statement;
 use tracing::info;
-use vector_search_validator_tests::ScyllaClusterExt;
-use vector_search_validator_tests::ScyllaNodeConfig;
 use vector_search_validator_tests::TestCase;
 use vector_search_validator_tests::TlsExt;
 use vector_search_validator_tests::VectorStoreNodeConfig;
@@ -40,7 +40,7 @@ async fn test_secondary_uri_works_correctly(actors: TestActors) {
             db_ip: actors.services_subnet.ip(DB_OCTET_1),
             primary_vs_uris: vec![vs_url.clone()],
             secondary_vs_uris: vec![],
-            args: vector_search_validator_tests::default_scylla_args(),
+            args: e2etest_scylla_cluster::default_scylla_args(),
             cert_path: Some(cert_path.clone()),
             key_path: Some(key_path.clone()),
             extra_config: None,
@@ -49,7 +49,7 @@ async fn test_secondary_uri_works_correctly(actors: TestActors) {
             db_ip: actors.services_subnet.ip(DB_OCTET_2),
             primary_vs_uris: vec![],
             secondary_vs_uris: vec![vs_url.clone()],
-            args: vector_search_validator_tests::default_scylla_args(),
+            args: e2etest_scylla_cluster::default_scylla_args(),
             cert_path: Some(cert_path.clone()),
             key_path: Some(key_path.clone()),
             extra_config: None,
@@ -58,7 +58,7 @@ async fn test_secondary_uri_works_correctly(actors: TestActors) {
             db_ip: actors.services_subnet.ip(DB_OCTET_3),
             primary_vs_uris: vec![],
             secondary_vs_uris: vec![vs_url.clone()],
-            args: vector_search_validator_tests::default_scylla_args(),
+            args: e2etest_scylla_cluster::default_scylla_args(),
             cert_path: Some(cert_path.clone()),
             key_path: Some(key_path.clone()),
             extra_config: None,
