+++ crates/validator/src/lib.rs
@@ -24,6 +24,8 @@ use async_backtrace::framed;
 use e2etest_dns::Dns;
 use e2etest_dns::DnsExt;
 use e2etest_firewall::Firewall;
+use e2etest_scylla_cluster::ScyllaCluster;
+use e2etest_scylla_cluster::ScyllaClusterExt;
 use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -34,8 +36,6 @@ use tracing_subscriber::EnvFilter;
 use tracing_subscriber::filter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::prelude::*;
-use vector_search_validator_tests::ScyllaCluster;
-use vector_search_validator_tests::ScyllaClusterExt;
 use vector_search_validator_tests::ScyllaProxyCluster;
 use vector_search_validator_tests::TestCase;
 use vector_search_validator_tests::Tls;
