+++ crates/validator/src/reconnect.rs
@@ -7,6 +7,7 @@ use crate::TestActors;
 use crate::common::*;
 use async_backtrace::framed;
 use e2etest_firewall::FirewallExt;
+use e2etest_scylla_cluster::ScyllaClusterExt;
 use scylla_proxy::Condition;
 use scylla_proxy::Reaction;
 use scylla_proxy::RequestReaction;
@@ -14,7 +15,6 @@ use scylla_proxy::RequestRule;
 use std::time::Duration;
 use tap::Pipe;
 use tracing::info;
-use vector_search_validator_tests::ScyllaClusterExt;
 use vector_search_validator_tests::ScyllaProxyClusterExt;
 use vector_search_validator_tests::TestCase;
 use vector_search_validator_tests::VectorStoreClusterExt;
